multimodal-ocr-system/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pdf2image>=1.16.3
PyPDF2>=3.0.1
easyocr>=1.7.0
pytesseract>=0.3.10
transformers>=4.35.0
torch>=2.0.0
spacy>=3.7.0
scikit-learn>=1.3.0
datasets>=2.14.0
mlflow>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.0.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
# Phase 2 additions
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.14.0